Adds dark-mode vs light-mode selection
This adds the mkdocs-material `scheme: preference` option in the
`mkdocs.yml` file to enable dark-mode vs light-mode selection based on
user preference.

The complication is that we overrode the colours for the search box
(because the alpha made them too dim) and now they need to switch
between white or black based on mode, so that's in the `tweaks.css`
file.