Corrects paths in updaters, adds force option
I'm not sure when the paths were changed to be /data/content instead of
just /data, but I'm pretty sure that wasn't announced. C_C

To add command-line options to the cron_update.sh script, I had to
restructure the main control flow a bit -- it didn't accept any
before.