Switches update detection mechanism to better one
The previous update mechanism used sed to grab a value out of json,
which is a little fragile. More importantly, the value it grabbed seemed
correct at the time it was written but later seemed to get stuck by
people editing files from within GitHub.

I'm not 100% sure on that, but in any case, this new mechanism detects
-- as far as I can tell from the GitHub documentation -- the date/time
of the most recent commit to the gh-pages branch.

Unfortunately we don't have a way to install packages on the webserver
and I didn't want to have to add a Python script to parse the JSON,
so I downloaded jq into ccspapp's home directory/bin to handle
it. It's not great and I'm tired of having to fight this excuse for a
PaaS.