Revert "Adds 'page last updated' feature"
This reverts commit 4c1c0d61fa8e2bd38af7276e39fa7a3d0d90f4c1.

The build script currently moves some things around in a way that means
the git data isn't available to mkdocs, which it needs to work out when
pages were last updated.

So this will need some more thought to work out what changes would be necessary in order to work.