Corrects links to images in Walkthrough
NB: these are *not* correct for the rendered page. mkdocs corrects them
at render time, so if you "fix" them to what they should be in the final
pages, mkdocs will complain that it cannot find them and spit out some
errors. Letting mkdocs correct them works just as well as writing them
for the the rendered page, *but* this way, mkdocs helps find actually missing things.

At least, that's the idea. Let's see if it works in practice.